+++ src/Firebase/Exception/RemoteConfigException.php
@@ -5,6 +5,7 @@ namespace Kreait\Firebase\Exception;
 use GuzzleHttp\Exception\RequestException;
 use Kreait\Firebase\Exception\RemoteConfig\OperationAborted;
 use Kreait\Firebase\Exception\RemoteConfig\PermissionDenied;
+use Kreait\Firebase\Exception\RemoteConfig\VersionMismatch;
 use Kreait\Firebase\Util\JSON;
 
 class RemoteConfigException extends \RuntimeException implements FirebaseException
@@ -12,6 +13,7 @@ class RemoteConfigException extends \RuntimeException implements FirebaseExcepti
     public static $errors = [
         PermissionDenied::IDENTIFER => PermissionDenied::class,
         OperationAborted::IDENTIFER => OperationAborted::class,
+        VersionMismatch::IDENTIFER => VersionMismatch::class,
     ];
 
     /**
@@ -31,7 +33,7 @@ class RemoteConfigException extends \RuntimeException implements FirebaseExcepti
 
         $candidates = array_filter(array_map(function ($key, $class) use ($message, $e) {
             return stripos($message, $key) !== false
-                ? new $class($e->getCode(), $e)
+                ? new $class($message, $e->getCode(), $e)
                 : null;
         }, array_keys(self::$errors), self::$errors));