+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 * Project IDs that cannot be used for database URIs directly are sanitized when configuring a Service Account ([#228](https://github.com/kreait/firebase-php/issues/228))
 
+## Non-breaking changes
+
+* When publishing an outdated Remote Config, Firebase previously returned an "OPERATION ABORTED" error,
+  resulting in a `Kreait\Firebase\Exception\RemoteConfig\OperationAborted` exception. Firebase has
+  changed the error to "VERSION MISMATCH", which now results in a `Kreait\Firebase\Exception\RemoteConfig\VersionMismatch` exception.
+  The new exception inherits from the old exception, so no code changes are required.
+* A `Kreait\Firebase\Exception\RemoteConfigException` now includes the full error as returned by the Firebase API.
+
 ## 4.13.1 - 2018-07-17
 
 ### Bugfixes